优化preview源码结构
